Remove `types` package
Move type definitions closer to where they are used.
